chat responsive UI optimization
